requirements.txt
openupgradelib==2.0.0
odoo-test-helper
